requirements.txt
alembic==1.4.2
autoenv==1.0.0
blinker==1.4
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
dnspython==1.16.0
dominate==2.5.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-DotEnv==0.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
passlib==1.7.2
protobuf==3.6.1
pycparser==2.20
psycopg2
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.2.1
